Tailorized "2005-12-01 23:05:43 by logg" Name change: Node --> Vertex
The class Node is now called Vertex (in accordance with a previous
discussion). It took a while but all occurences of "Node" in the code
should now be replaced with "Vertex" (and similar for other variants
such as "Nodes", "node", "NODE"). The XML data format and all data
files have also been updated.

I have not changed the name of all variables of type Node named 'n' or
'n0', 'n1' etc to 'v*' though...