Tailorized "2005-12-06 13:54:27 by garth" NLS:
Change from PETSc nonlinear solvers to own Newton methods. It looks a lot cleaner and simpler. It is derived from KrylovSolver, so parameters for the linear solver are set in the usual way.

There are two interfaces to the nonlinear solver.

1. The simple version (see src/demo/nls/poisson/):

NewtonSolver newtonsolver;
newtonsolver.solve(a, L, bc, mesh, x);


2. More sophisicated version (see src/demo/nls/nonlinearpoisson):

MyNonlinearFunction nonlinear_function(a, L, mesh, bc);

NewtonSolver newtonsolver;
newtonsolver.solve(nonlinear_function, x);

The user defines the NonlinearFunction. It returns the vector F(u)=0 and the Jacobian for the particular problem.