Remove f0tmp in MultiAdaptiveTimeSlab.
Instead of putting f at right end-point in f0tmp and then
copying to f0 if the slab is accepted, the values are now
always put into f0 and the values recomputed in the case
when the slab is not accepted (which should be just a few
cases). Seems to work and the performance is the same.