Fix compilation with --disable-petsc. (Should work now Åsmund.)
A number of the demos don't compile without PETSc, but all libraries compile
fine. We should remove the dependence on PETSc and work through GenericMatrix
as much as possible. We should also consider doing a typedef of DenseMatrix to
Matrix when running with --disable-petsc (instead of SparseMatrix).

We should set up a compilation bot on fenics.org to create nightly builds
for various combinations of configurations. Otherwise we'll run into trouble
with flags that we rarely use (like --disable-petsc).